+++ client/src/components/LoginButton.tsx
@@ -1,20 +1,38 @@
-import { ReactElement } from 'react';
-import styled from 'styled-components';
-import Login from '@images/user.svg';
+import { ReactProps } from '@type/props';
+import { ReactElement, useState } from 'react';
+import { useQueryClient } from 'react-query';
+import { toast } from 'react-toastify';
+import LoginModal from './LoginModal';
 
-const StyledButton = styled.button`
-  background-color: transparent;
-  padding: auto;
-  height: 84px;
-  width: 84px;
-  border: none;
-`;
+function LoginButton({ children }: ReactProps<{}>): ReactElement {
+  const [showModal, setShowModal] = useState<boolean>(false);
+  const queryClient = useQueryClient();
 
-function LoginButton({ onClick }: { onClick: () => void }): ReactElement {
+  const handleOnClick = (): void => {
+    if (queryClient.getQueryData('user') === undefined) setShowModal(true);
+    else {
+      toast.success('로그아웃되었습니다.');
+      queryClient.setQueryData('user', undefined);
+    }
+  };
+
   return (
-    <StyledButton onClick={onClick}>
-      <img src={Login} />
-    </StyledButton>
+    <div>
+      <div
+        onClick={() => {
+          handleOnClick();
+        }}
+      >
+        {children}
+      </div>
+      {showModal && (
+        <LoginModal
+          onCancel={() => {
+            setShowModal(false);
+          }}
+        />
+      )}
+    </div>
   );
 }
